+++ tests/__init__.py
@@ -19,6 +19,22 @@ import os
 import unittest
 
 
+def get_path(code_file):
+    """Return relative or absolute path to given code file.
+
+    During Debian package build, the current directory is
+    changed to .pybuild/cpython3_*/build and the root directory
+    is stored in the OLDPWD environment variable.
+    """
+    if os.path.exists(code_file):
+        return code_file
+    # The alternative path is needed for Debian's pybuild
+    alternative = os.path.join(os.environ.get("OLDPWD", ""), code_file)
+    if os.path.exists(alternative):
+        return alternative
+    return code_file
+
+
 def get_source_files():
     """Return a list of sources files/directories (to check with flake8/pylint)."""
     scripts = ["bdebstrap"]
@@ -27,18 +43,14 @@ def get_source_files():
 
     files = []
     for code_file in scripts + modules + py_files:
-        is_script = code_file in scripts
-        if not os.path.exists(code_file):  # pragma: no cover
-            # The alternative path is needed for Debian's pybuild
-            alternative = os.path.join(os.environ.get("OLDPWD", ""), code_file)
-            code_file = alternative if os.path.exists(alternative) else code_file
-        if is_script:
-            with open(code_file, "rb") as script_file:
+        code_path = get_path(code_file)
+        if code_file in scripts:
+            with open(code_path, "rb") as script_file:
                 shebang = script_file.readline().decode("utf-8")
             if "python" in shebang:
-                files.append(code_file)
+                files.append(code_path)
         else:
-            files.append(code_file)
+            files.append(code_path)
     return files
 
 
@@ -55,3 +67,32 @@ def unittest_verbosity():
             return self.verbosity
         frame = frame.f_back
     return 0  # pragma: no cover
+
+
+class TestGetPath(unittest.TestCase):
+    """Unittests for get_path function."""
+
+    def test_get_path_exists(self):
+        """Test get_path(__file__)."""
+        path = get_path(__file__)
+        self.assertEqual(path, __file__)
+
+    def test_get_path_missing(self):
+        """Test non-existing file for get_path()."""
+        path = get_path(__file__ + "non-existing")
+        self.assertEqual(path, __file__ + "non-existing")
+
+    def test_get_path_pybuild(self):
+        """Test changing current directory before calling get_path()."""
+        relpath = os.path.relpath(__file__)
+        oldpwd = os.environ.get("OLDPWD")
+        curdir = os.getcwd()
+        try:
+            os.environ["OLDPWD"] = curdir
+            os.chdir("/")
+            path = get_path(relpath)
+        finally:
+            os.chdir(curdir)
+            if oldpwd is not None:
+                os.environ["OLDPWD"] = oldpwd
+        self.assertEqual(os.path.normpath(path), __file__)